Move cert compression to the cert module
agent.rs is already big enough, thanks.

Also, clean up some imports in the process.  Notably, we've been
importing `TryFrom`, which was added to the rust prelude in Edition
2021.

This is not API-compatible, because stuff moved.  It can be made
compatible by adding a `pub use`, but I'm currently of the view that it
would be better if it didn't go that way.
