+++ src/hash.rs
@@ -4,8 +4,6 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-use std::convert::TryFrom as _;
-
 use crate::{
     Error,
     err::IntoResult as _,
